requirements.txt
python-dotenv>=0.21.0
streamlit>=1.25.0
openai>=1.0.0
